requirements.txt
beautifulsoup4
requests
html5lib
PyYAML
